+++ changelog.md
@@ -1,2 +1,8 @@
-# 1.0.0
+# 1.0.1
 - initial release
+
+# 1.0.2
+- fix game crashes after updating create menu
+
+# 1.0.3
+- add new contrast button sprites